% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom-plot-function.R
\name{custom_plot_gadget}
\alias{custom_plot_gadget}
\title{Shiny Gadget for Plot Customization}
\usage{
custom_plot_gadget(data = mtcars)
}
\arguments{
\item{data}{A data frame to use for the plot. Defaults to `mtcars`.}
}
\value{
None. Opens a Shiny app.
}
\description{
Launches a Shiny gadget to customize plots interactively, including the ability to set a custom title.
}
\examples{
if (interactive()) custom_plot_gadget(data = iris)
}
