% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-function.R
\name{palette_vibrant}
\alias{palette_vibrant}
\title{Vibrant Color Palette}
\usage{
palette_vibrant()
}
\value{
A vector of hex color codes.
}
\description{
A custom color palette with vibrant and eye-catching colors for use in plots.
}
\details{
This palette can be applied using `scale_color_manual()` for categorical variables
or `scale_color_gradient()` for continuous variables to add vibrant colors to ggplot visualizations.
}
\examples{
ggplot(mtcars, aes(mpg, wt)) +
  geom_point(aes(color = factor(cyl))) +
  scale_color_manual(values = palette_vibrant())

ggplot(mtcars, aes(mpg, wt)) +
  geom_point(aes(color = mpg)) +
  scale_color_gradientn(colors = palette_vibrant())
}
